JAMES-2521 Move JPA migrator utils as a separate tool project
